requirements.txt
discord.py
psutil
